Make my module a function.
It appears to me that a module can be just a plain attribute
set, or it can be a function.

We want it to be a function because, if we eventually move
it to another file, we won't be able to close over the `pkgs`
defined in the let block of the flake top-level, which means
we need to pass it down. We could probably do this ourselves,
but home manager will pass it for us.