Install antidote with nix, but plugins manually
Summary:

For reasons I don't understand, the plugins don't work on my linux
machine; I wind up with a functional antidote but no plugins.

The workaround is to go hybrid: rely on the flake to install and enable
antidote, but handle the bundling and loading manually.

I forced the bundling to use ssh; by default it tries to use https, and
I get prompted for login information which is really not a great
experience.