Tried to update my neovim to 0.11, failed
I seem to have bumped home-manager to 24.11 which is good, and neovim
is bumped to an 0.10.x version which is also good.

But I expected neovim-nightly-overlay to give me a current neovim
and it did not, 0.11 was released in March and I'm still getting
an older version :/

I'm not sure what's going on yet, but this is at least progress.